Move watchasset to global
